requirements.txt
aiosqlite==0.17.0
asgiref==3.4.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
Django==2.1.12
fastapi==0.68.1
h11==0.12.0
idna==3.2
iso8601==0.1.16
limits==1.5.1
lxml==4.6.3
MouseInfo==0.1.3
passlib==1.7.4
Pillow==8.2.0
PyAutoGUI==0.9.52
pycparser==2.20
pydantic==1.8.2
PyGetWindow==0.0.9
PyJWT==2.1.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypika-tortoise==0.1.1
PyRect==0.1.4
PyScreeze==0.1.27
python-multipart==0.0.5
PyTweening==1.0.3
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
slowapi==0.1.5
soupsieve==2.2.1
starlette==0.14.2
tortoise-orm==0.17.7
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
